requirements.txt
alabaster==1.0.0
babel==2.17.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
docutils==0.21.2
idna==3.11
imagesize==1.4.1
indic-nlp-library==0.92
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mlmorph==1.4.0
Morfessor==2.0.6
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.10.23
requests==2.32.5
roman-numerals-py==3.1.0
sfst==1.8.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
